feat Gregory [Final cart bug fixes]
