--- E-Commerce/Modules/Cart/CartScreen/CartView.swift
@@ -1,12 +1,0 @@
-//
-//  CartView.swift
-//  E-Commerce
-//
-//  Created by Alexander Bokhulenkov on 03.03.2025.
-//
-
-import UIKit
-
-class CartView: UIView {
-    
-}
